refactoring the code to make it scalable
