+++ ToDoApp/routers/todos.py
@@ -1,0 +1,79 @@
+from typing import Annotated
+
+from fastapi import APIRouter, Depends, HTTPException, Path
+from pydantic import BaseModel, Field
+from sqlalchemy.orm import Session
+from starlette import status
+
+from models import Todos
+from database import SessionLocal
+
+router = APIRouter()
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+db_dependency = Annotated[Session, Depends(get_db)]
+
+
+class TodoRequest(BaseModel):
+    title: str = Field(min_length=3)
+    description: str = Field(min_length=3, max_length=100)
+    priority: int = Field(gt=0, lt=6)
+    complete: bool
+
+
+@router.get("/", status_code=status.HTTP_200_OK)
+async def read_all(db: db_dependency):
+    return db.query(Todos).all()
+
+
+@router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
+async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if todo_model:
+        return todo_model
+    else:
+        raise HTTPException(status_code=404, detail=f"todo_id {todo_id}, not found.")
+
+
+@router.post("/todo", status_code=status.HTTP_201_CREATED)
+async def create_todo(db: db_dependency, todo_req: TodoRequest):
+    todo_model = Todos(**todo_req.dict())
+
+    db.add(todo_model)
+    db.commit()
+
+
+@router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def update_todo(db: db_dependency, todo_req: TodoRequest, todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+
+    if not todo_model:
+        raise HTTPException(status_code=404, detail=f"todo_id {todo_id}, not found.")
+
+    todo_model.title = todo_req.title
+    todo_model.description = todo_req.description
+    todo_model.priority = todo_req.priority
+    todo_model.complete = todo_req.complete
+
+    db.add(todo_model)
+    db.commit()
+
+
+@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+
+    if not todo_model:
+        raise HTTPException(status_code=404, detail=f"todo_id {todo_id}, not found.")
+
+    db.query(Todos).filter(Todos.id == todo_id).delete()
+
+    db.commit()
